fix error interface name
