+++ ontology/smart_contract/neo_contract/invoke_function.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
