+++ ontology/utils/contract_event_parser.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
