--- ontology/utils/test_deserialize_hex.py
@@ -1,3 +1,0 @@
-from unittest import TestCase
-
-
